+++ confmap/internal/e2e/types_test.go
@@ -11,6 +11,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
+	"go.opentelemetry.io/collector/config/configopaque"
 	"go.opentelemetry.io/collector/confmap"
 	"go.opentelemetry.io/collector/confmap/provider/envprovider"
 	"go.opentelemetry.io/collector/confmap/provider/fileprovider"
@@ -421,7 +422,7 @@ func TestIssue10787(t *testing.T) {
 	assert.Equal(t, conf.ToStringMap(),
 		map[string]any{
 			"exporters": map[string]any{
-				"logging": map[string]any{
+				"debug": map[string]any{
 					"verbosity": "detailed",
 				},
 			},
@@ -443,7 +444,7 @@ func TestIssue10787(t *testing.T) {
 			"service": map[string]any{
 				"pipelines": map[string]any{
 					"traces": map[string]any{
-						"exporters":  []any{"logging"},
+						"exporters":  []any{"debug"},
 						"processors": []any{"batch"},
 						"receivers":  []any{"otlp"},
 					},
@@ -461,16 +462,16 @@ func TestIssue10787(t *testing.T) {
 func TestStructMappingIssue10787(t *testing.T) {
 	resolver := NewResolver(t, "types_expand.yaml")
 	t.Setenv("ENV", `# this is a comment
-logging:
+debug:
   verbosity: detailed`)
 	conf, err := resolver.Resolve(context.Background())
 	require.NoError(t, err)
 
-	type Logging struct {
+	type Debug struct {
 		Verbosity string `mapstructure:"verbosity"`
 	}
 	type Exporters struct {
-		Logging Logging `mapstructure:"logging"`
+		Debug Debug `mapstructure:"debug"`
 	}
 	type Target struct {
 		Field Exporters `mapstructure:"field"`
@@ -481,7 +482,7 @@ logging:
 	require.NoError(t, err)
 	require.Equal(t,
 		Target{Field: Exporters{
-			Logging: Logging{
+			Debug: Debug{
 				Verbosity: "detailed",
 			},
 		}},
@@ -494,7 +495,7 @@ logging:
 	err = confStr.Unmarshal(&cfgStr)
 	require.NoError(t, err)
 	require.Equal(t, `# this is a comment
-logging:
+debug:
   verbosity: detailed`,
 		cfgStr.Field,
 	)
@@ -504,16 +505,16 @@ func TestStructMappingIssue10787_ExpandComment(t *testing.T) {
 	resolver := NewResolver(t, "types_expand.yaml")
 	t.Setenv("EXPAND_ME", "an expanded env var")
 	t.Setenv("ENV", `# this is a comment with ${EXPAND_ME}
-logging:
+debug:
   verbosity: detailed`)
 	conf, err := resolver.Resolve(context.Background())
 	require.NoError(t, err)
 
-	type Logging struct {
+	type Debug struct {
 		Verbosity string `mapstructure:"verbosity"`
 	}
 	type Exporters struct {
-		Logging Logging `mapstructure:"logging"`
+		Debug Debug `mapstructure:"debug"`
 	}
 	type Target struct {
 		Field Exporters `mapstructure:"field"`
@@ -524,7 +525,7 @@ logging:
 	require.NoError(t, err)
 	require.Equal(t,
 		Target{Field: Exporters{
-			Logging: Logging{
+			Debug: Debug{
 				Verbosity: "detailed",
 			},
 		}},
@@ -537,7 +538,7 @@ logging:
 	err = confStr.Unmarshal(&cfgStr)
 	require.NoError(t, err)
 	require.Equal(t, `# this is a comment with an expanded env var
-logging:
+debug:
   verbosity: detailed`,
 		cfgStr.Field,
 	)
@@ -584,4 +585,74 @@ func TestIndirectSliceEnvVar(t *testing.T) {
 	assert.Equal(t, collectorConf.Exporters.OTLP.Endpoint, "localhost:4317")
 	assert.Equal(t, collectorConf.Service.Pipelines.Logs.Receivers, []string{"nop", "otlp"})
 	assert.Equal(t, collectorConf.Service.Pipelines.Logs.Exporters, []string{"otlp", "nop"})
+}
+
+func TestIssue10937_MapType(t *testing.T) {
+	t.Setenv("VALUE", "1234")
+
+	resolver := NewResolver(t, "types_map.yaml")
+	conf, err := resolver.Resolve(context.Background())
+	require.NoError(t, err)
+
+	var cfg TargetConfig[map[string]configopaque.String]
+	err = conf.Unmarshal(&cfg)
+	require.NoError(t, err)
+	require.Equal(t, map[string]configopaque.String{"key": "1234"}, cfg.Field)
+}
+
+func TestIssue10937_ArrayType(t *testing.T) {
+	t.Setenv("VALUE", "1234")
+
+	resolver := NewResolver(t, "types_slice.yaml")
+	conf, err := resolver.Resolve(context.Background())
+	require.NoError(t, err)
+
+	var cfgStrSlice TargetConfig[[]string]
+	err = conf.Unmarshal(&cfgStrSlice)
+	require.NoError(t, err)
+	require.Equal(t, []string{"1234"}, cfgStrSlice.Field)
+
+	var cfgStrArray TargetConfig[[1]string]
+	err = conf.Unmarshal(&cfgStrArray)
+	require.NoError(t, err)
+	require.Equal(t, [1]string{"1234"}, cfgStrArray.Field)
+
+	var cfgAnySlice TargetConfig[[]any]
+	err = conf.Unmarshal(&cfgAnySlice)
+	require.NoError(t, err)
+	require.Equal(t, []any{1234}, cfgAnySlice.Field)
+
+	var cfgAnyArray TargetConfig[[1]any]
+	err = conf.Unmarshal(&cfgAnyArray)
+	require.NoError(t, err)
+	require.Equal(t, [1]any{1234}, cfgAnyArray.Field)
+}
+
+func TestIssue10937_ComplexType(t *testing.T) {
+	t.Setenv("VALUE", "1234")
+
+	resolver := NewResolver(t, "types_complex.yaml")
+	conf, err := resolver.Resolve(context.Background())
+	require.NoError(t, err)
+
+	var cfgStringy TargetConfig[[]map[string][]string]
+	err = conf.Unmarshal(&cfgStringy)
+	require.NoError(t, err)
+	require.Equal(t, []map[string][]string{{"key": {"1234"}}}, cfgStringy.Field)
+
+	var cfgNotStringy TargetConfig[[]map[string][]any]
+	err = conf.Unmarshal(&cfgNotStringy)
+	require.NoError(t, err)
+	require.Equal(t, []map[string][]any{{"key": {1234}}}, cfgNotStringy.Field)
+}
+
+func TestIssue10949_UnsetVar(t *testing.T) {
+	resolver := NewResolver(t, "types_expand.yaml")
+	conf, err := resolver.Resolve(context.Background())
+	require.NoError(t, err)
+
+	var cfg TargetConfig[int]
+	err = conf.Unmarshal(&cfg)
+	require.NoError(t, err)
+	require.Equal(t, 0, cfg.Field)
 }